% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapImporter.R
\name{redcap.importer}
\alias{redcap.importer}
\title{REDCap importer function
                                            
This function takes patient data from HIVE Enrollment project on REDCap and processes it, transforming the data into that described by the REDCap import template}
\usage{

  redcap.importer(hive.enroll.file = "U:/EPID/Monto/HIVE_2016_2017/REDCap/Data Exports/HIVE enrollment export update.csv",
  hive.study.file = "U:/EPID/Monto/HIVE_2016_2017/REDCap/Data Exports/HIVE study import update.csv",
  arm = 2)
}
\arguments{
\item{hive.enroll.file}{Data exported from the HIVE Enrollment project on REDCap as a file path ending in .csv, enclosed in quotes, defaults to "U:/EPID/Monto/HIVE_2016_2017/REDCap/Data Exports/HIVE enrollment export update.csv".}

\item{hive.study.file}{The output file name as a file path enclosed in quotes. Defaults to "U:/EPID/Monto/HIVE_2016_2017/REDCap/Data Exports/HIVE study import update.csv".}

\item{arm}{Specify the arm of the study. the function accepts 1, 2 and 3 for re-enrollees, new HIVE, and pediatric HIVE, respectively. Defaults to 2.}
}
\description{
REDCap importer function
                                            
This function takes patient data from HIVE Enrollment project on REDCap and processes it, transforming the data into that described by the REDCap import template
}

